User Interface Design Project
This project was created as a final for our User Interface Design class. I was one of a team of three who worked on it. We didn't have much experience using CSS, HTML, and JavaScript going in, but I took some time after the class to learn REACT. Over the course of the class, we would perform usability tests with friends in the Central Washington University music program, and made revisions based on how they used it. The project could definitely use polishing, but the experience we received in usability testing is incredibly valuable to us.